+++ dev_notes
@@ -10,14 +10,17 @@ DEV NOTES
 
 Feature branch:
 
-speech_recognition_feature
+2. speech_recognition_feature
 
-1 -> implement this (https://developer.mozilla.org/en-US/docs/Web/API/MediaStream_Recording_API/Using_the_MediaStream_Recording_API) example of a web dictaphone
+1. implement this (https://developer.mozilla.org/en-US/docs/Web/API/MediaStream_Recording_API/Using_the_MediaStream_Recording_API) example of a web dictaphone
+	- github link: https://github.com/mdn/web-dictaphone
 
-2 -> implement the basic speech_recognition library example
+1.5. send the audio blob to the flask server backend. save the blob as an audio file through the backend (to confirm that the correct blob was transmitted)
 
-3 -> use the recorded audio sample from 1 and use it to make inference from 2
+2. implement the basic speech_recognition library example
 
-4 -> cleanup useless/excess features code
+3. use the recorded audio sample from 1 and use it to make inference from 2
 
-5 -> [fancy UI feature; not core functionality] add realtime visualization of audio levels >> https://speckyboy.com/audio-visualization-code-snippets/ 
+4. cleanup useless/excess features code
+
+5. [fancy UI feature; not core functionality] add realtime visualization of audio levels >> https://speckyboy.com/audio-visualization-code-snippets/
